+++ src/daemon/handlers/find-match-resolution.ts
@@ -5,11 +5,9 @@ import {
 } from '@agent-device/selectors';
 import { listSelectorPipelineMatches } from '../../core/selector-pipeline.ts';
 import { SELECTOR_PIPELINE_POLICIES } from '../../core/selector-pipeline-policy.ts';
-import { centerOfRect, type SnapshotState } from '@agent-device/kernel/snapshot';
-import {
-  isRootInteractionContainer,
-  resolveActionableTouchResolution,
-} from '../../core/interaction-targeting.ts';
+import type { SnapshotState } from '@agent-device/kernel/snapshot';
+import { isRootInteractionContainer } from '../../core/interaction-targeting.ts';
+import { preferOnscreenMatches } from './find-match-ranking.ts';
 import { formatSnapshotLine } from '../../snapshot/snapshot-lines.ts';
 import type { ElementMatchCandidateDetails } from '../../utils/error-candidates.ts';
 import type { DaemonRequest, DaemonResponse, SessionState } from '../types.ts';
@@ -91,71 +89,6 @@ function narrowMultipleMatches(
   return null;
 }
 
-function preferOnscreenMatches(
-  matches: SnapshotState['nodes'],
-  nodes: SnapshotState['nodes'],
-): SnapshotState['nodes'] {
-  const viewport = nodes[0]?.rect;
-  if (!viewport) return matches;
-  const onscreen = matches.filter((node) => {
-    if (!node.rect) return false;
-    const center = centerOfRect(node.rect);
-    return (
-      center.x >= viewport.x &&
-      center.x <= viewport.x + viewport.width &&
-      center.y >= viewport.y &&
-      center.y <= viewport.y + viewport.height
-    );
-  });
-  return rankInteractiveMatches(onscreen.length > 0 ? onscreen : matches, nodes);
-}
-
-function rankInteractiveMatches(
-  matches: SnapshotState['nodes'],
-  nodes: SnapshotState['nodes'],
-): SnapshotState['nodes'] {
-  if (matches.length < 2) return matches;
-  return matches
-    .map((node, index) => ({ node, index, score: interactiveMatchScore(node, nodes) }))
-    .sort((left, right) => {
-      if (right.score !== left.score) return right.score - left.score;
-      return rectArea(left.node) - rectArea(right.node) || left.index - right.index;
-    })
-    .map((entry) => entry.node);
-}
-
-function interactiveMatchScore(
-  node: SnapshotState['nodes'][number],
-  nodes: SnapshotState['nodes'],
-): number {
-  const resolution = resolveActionableTouchResolution(nodes, node);
-  if (resolution.reason === 'covered') return 0;
-  const resolved = resolvedTouchScore(resolution, nodes[0]);
-  if (resolved > 0) return resolved;
-  if (node.hittable && node.rect && !isRootInteractionContainer(node, nodes[0])) return 3;
-  return node.rect ? 1 : 0;
-}
-
-function resolvedTouchScore(
-  resolution: ReturnType<typeof resolveActionableTouchResolution>,
-  root: SnapshotState['nodes'][number] | undefined,
-): number {
-  if (!resolution.node.rect) return 0;
-  if (resolution.reason === 'semantic-target' || resolution.reason === 'same-rect-descendant') {
-    return 4;
-  }
-  if (
-    resolution.reason === 'hittable-ancestor' &&
-    !isRootInteractionContainer(resolution.node, root)
-  ) {
-    return 2;
-  }
-  return 0;
-}
-
-function rectArea(node: SnapshotState['nodes'][number]): number {
-  return node.rect ? node.rect.width * node.rect.height : Number.POSITIVE_INFINITY;
-}
 // #1597: an agent reading an ambiguous-match error must be able to act on the
 // right @ref immediately, without a follow-up snapshot round trip. Candidate
 // lines reuse the exact snapshot-line renderer (`formatSnapshotLine`) so a
